+++ CHANGES.md
@@ -1,7 +1,12 @@
 ## unreleased
 
+- Add new `name_ns_str`, `namespace_str` and `prefix_str` methods to access
+  name information.
+
 - Changed the processing instruction API to not use `get_target` and
   `get_value` but instead to use `target()` and `value()`.
+
+- Make `NameId`, `NamespaceId` and `PrefixId` public.
 
 - Documentation tweaks.